requirements.txt
django>=3.2,<4.0
pillow>=8.2.0,<8.3
graphene-django>=2.15.0,<2.16
graphene-file-upload>=1.3.0,<1.4.0
django-filter>=2.4.0,<2.5
django-cors-headers>=3.7.0,<3.8
black>=21.7b0,<21.8b0
psycopg2>=2.8.6,<2.9
gunicorn>=20.1.0,<20.2
django-on-heroku>=1.1.2,<1.2
django-cloudinary-storage>=0.3.0
django-graphql-jwt